# api/swagger.yaml
basePath: /damselfly
definitions:
  handler.CloudModelRespInfo:
    properties:
      cloudInfraModel:
        $ref: '#/definitions/model.TbMciDynamicReq'
      cloudModelVersion:
        type: string
      createTime:
        type: string
      csp:
        type: string
      description:
        type: string
      id:
        type: string
      isCloudModel:
        type: boolean
      isInitUserModel:
        type: boolean
      isTargetModel:
        type: boolean
      region:
        type: string
      updateTime:
        type: string
      userId:
        type: string
      userModelName:
        type: string
      userModelVersion:
        type: string
      zone:
        type: string
    required:
    - cloudInfraModel
    type: object
  handler.CreateCloudModelReq:
    properties:
      cloudInfraModel:
        $ref: '#/definitions/model.TbMciDynamicReq'
      csp:
        type: string
      description:
        type: string
      isInitUserModel:
        type: boolean
      isTargetModel:
        type: boolean
      region:
        type: string
      userId:
        type: string
      userModelName:
        type: string
      userModelVersion:
        type: string
      zone:
        type: string
    required:
    - cloudInfraModel
    type: object
  handler.CreateCloudModelResp:
    properties:
      cloudInfraModel:
        $ref: '#/definitions/model.TbMciDynamicReq'
      cloudModelVersion:
        type: string
      createTime:
        type: string
      csp:
        type: string
      description:
        type: string
      id:
        type: string
      isCloudModel:
        type: boolean
      isInitUserModel:
        type: boolean
      isTargetModel:
        type: boolean
      region:
        type: string
      updateTime:
        type: string
      userId:
        type: string
      userModelName:
        type: string
      userModelVersion:
        type: string
      zone:
        type: string
    required:
    - cloudInfraModel
    type: object
  handler.CreateOnPremModelReq:
    properties:
      description:
        type: string
      isInitUserModel:
        type: boolean
      onpremiseInfraModel:
        $ref: '#/definitions/inframodel.OnpremInfra'
      userId:
        type: string
      userModelName:
        type: string
      userModelVersion:
        type: string
    required:
    - onpremiseInfraModel
    type: object
  handler.CreateOnPremModelResp:
    properties:
      createTime:
        type: string
      description:
        type: string
      id:
        type: string
      isCloudModel:
        type: boolean
      isInitUserModel:
        type: boolean
      isTargetModel:
        type: boolean
      onpremModelVersion:
        type: string
      onpremiseInfraModel:
        $ref: '#/definitions/inframodel.OnpremInfra'
      updateTime:
        type: string
      userId:
        type: string
      userModelName:
        type: string
      userModelVersion:
        type: string
    required:
    - onpremiseInfraModel
    type: object
  handler.GetCloudModelResp:
    properties:
      cloudInfraModel:
        $ref: '#/definitions/model.TbMciDynamicReq'
      cloudModelVersion:
        type: string
      createTime:
        type: string
      csp:
        type: string
      description:
        type: string
      id:
        type: string
      isCloudModel:
        type: boolean
      isInitUserModel:
        type: boolean
      isTargetModel:
        type: boolean
      region:
        type: string
      updateTime:
        type: string
      userId:
        type: string
      userModelName:
        type: string
      userModelVersion:
        type: string
      zone:
        type: string
    required:
    - cloudInfraModel
    type: object
  handler.GetCloudModelsResp:
    properties:
      models:
        items:
          $ref: '#/definitions/handler.CloudModelRespInfo'
        type: array
    type: object
  handler.GetModelsResp:
    properties:
      models:
        items:
          $ref: '#/definitions/handler.ModelRespInfo'
        type: array
    type: object
  handler.GetModelsVersionResp:
    properties:
      modelsVersion:
        $ref: '#/definitions/handler.ModelsVersionRespInfo'
    type: object
  handler.GetOnPremModelResp:
    properties:
      createTime:
        type: string
      description:
        type: string
      id:
        type: string
      isCloudModel:
        type: boolean
      isInitUserModel:
        type: boolean
      isTargetModel:
        type: boolean
      onpremModelVersion:
        type: string
      onpremiseInfraModel:
        $ref: '#/definitions/inframodel.OnpremInfra'
      updateTime:
        type: string
      userId:
        type: string
      userModelName:
        type: string
      userModelVersion:
        type: string
    required:
    - onpremiseInfraModel
    type: object
  handler.GetOnPremModelsResp:
    properties:
      models:
        items:
          $ref: '#/definitions/handler.OnPremModelRespInfo'
        type: array
    type: object
  handler.ModelRespInfo:
    properties:
      cloudInfraModel:
        $ref: '#/definitions/model.TbMciDynamicReq'
      cloudModelVersion:
        type: string
      createTime:
        type: string
      csp:
        type: string
      description:
        type: string
      id:
        type: string
      isCloudModel:
        type: boolean
      isInitUserModel:
        type: boolean
      isTargetModel:
        type: boolean
      onpremModelVersion:
        type: string
      onpremiseInfraModel:
        $ref: '#/definitions/inframodel.OnpremInfra'
      region:
        type: string
      updateTime:
        type: string
      userId:
        type: string
      userModelName:
        type: string
      userModelVersion:
        type: string
      zone:
        type: string
    required:
    - cloudInfraModel
    - onpremiseInfraModel
    type: object
  handler.ModelsVersionRespInfo:
    properties:
      cloudModelVersion:
        type: string
      onpremModelVersion:
        type: string
    type: object
  handler.OnPremModelRespInfo:
    properties:
      createTime:
        type: string
      description:
        type: string
      id:
        type: string
      isCloudModel:
        type: boolean
      isInitUserModel:
        type: boolean
      isTargetModel:
        type: boolean
      onpremModelVersion:
        type: string
      onpremiseInfraModel:
        $ref: '#/definitions/inframodel.OnpremInfra'
      updateTime:
        type: string
      userId:
        type: string
      userModelName:
        type: string
      userModelVersion:
        type: string
    required:
    - onpremiseInfraModel
    type: object
  handler.ResHTTPVersion:
    properties:
      message:
        type: string
    type: object
  handler.ResReadyz:
    properties:
      message:
        type: string
    type: object
  handler.UpdateCloudModelReq:
    properties:
      cloudInfraModel:
        $ref: '#/definitions/model.TbMciDynamicReq'
      csp:
        type: string
      description:
        type: string
      isInitUserModel:
        type: boolean
      isTargetModel:
        type: boolean
      region:
        type: string
      userId:
        type: string
      userModelName:
        type: string
      userModelVersion:
        type: string
      zone:
        type: string
    required:
    - cloudInfraModel
    type: object
  handler.UpdateCloudModelResp:
    properties:
      cloudInfraModel:
        $ref: '#/definitions/model.TbMciDynamicReq'
      cloudModelVersion:
        type: string
      createTime:
        type: string
      csp:
        type: string
      description:
        type: string
      id:
        type: string
      isCloudModel:
        type: boolean
      isInitUserModel:
        type: boolean
      isTargetModel:
        type: boolean
      region:
        type: string
      updateTime:
        type: string
      userId:
        type: string
      userModelName:
        type: string
      userModelVersion:
        type: string
      zone:
        type: string
    required:
    - cloudInfraModel
    type: object
  handler.UpdateOnPremModelReq:
    properties:
      description:
        type: string
      isInitUserModel:
        type: boolean
      onpremiseInfraModel:
        $ref: '#/definitions/inframodel.OnpremInfra'
      userId:
        type: string
      userModelName:
        type: string
      userModelVersion:
        type: string
    required:
    - onpremiseInfraModel
    type: object
  handler.UpdateOnPremModelResp:
    properties:
      createTime:
        type: string
      description:
        type: string
      id:
        type: string
      isCloudModel:
        type: boolean
      isInitUserModel:
        type: boolean
      isTargetModel:
        type: boolean
      onpremModelVersion:
        type: string
      onpremiseInfraModel:
        $ref: '#/definitions/inframodel.OnpremInfra'
      updateTime:
        type: string
      userId:
        type: string
      userModelName:
        type: string
      userModelVersion:
        type: string
    required:
    - onpremiseInfraModel
    type: object
  inframodel.CpuProperty:
    properties:
      architecture:
        example: x86_64
        type: string
      cores:
        description: Number of physical cores per CPU
        example: 18
        type: integer
      cpus:
        description: Number of physical CPUs (sockets)
        example: 2
        type: integer
      maxSpeed:
        description: Maximum speed in GHz
        example: 3.6
        type: number
      model:
        example: Intel(R) Xeon(R) Gold 6140 CPU @ 2.30GHz
        type: string
      threads:
        description: Number of logical CPUs (threads) per CPU with hyper-threading
          enabled
        example: 36
        type: integer
      vendor:
        example: GenuineIntel
        type: string
    required:
    - cores
    - cpus
    - threads
    type: object
  inframodel.DiskProperty:
    properties:
      available:
        description: Unit GiB
        type: integer
      label:
        type: string
      totalSize:
        description: Unit GiB
        example: 1024
        type: integer
      type:
        description: SSD, HDD
        example: SSD
        type: string
      used:
        description: Unit GiB
        type: integer
    required:
    - label
    - totalSize
    - type
    type: object
  inframodel.MemoryProperty:
    properties:
      available:
        description: Unit GiB
        type: integer
      totalSize:
        description: Unit GiB
        example: 128
        type: integer
      type:
        example: DDR4
        type: string
      used:
        description: Unit GiB
        type: integer
    required:
    - totalSize
    - type
    type: object
  inframodel.NetworkInterfaceProperty:
    properties:
      ipv4CidrBlocks:
        description: IPv4 address with prefix length (e.g., 192.168.0.21/24), instead
          of inet addr, Bcast, and Mask
        items:
          type: string
        type: array
      ipv6CidrBlocks:
        description: IPv6 address with prefix length (e.g., "2001:db8::1/64")
        items:
          type: string
        type: array
      macAddress:
        description: MAC address
        type: string
      mtu:
        description: Maximum Transmission Unit (MTU) in bytes
        type: integer
      name:
        description: Interface name (e.g., eth0, ens01, enp0s3)
        type: string
      state:
        description: Interface state (e.g., UP, DOWN)
        type: string
    required:
    - name
    type: object
  inframodel.NetworkProperty:
    properties:
      ipv4Networks:
        example:
        - 172.26.240.0/20
        items:
          type: string
        type: array
      ipv6Networks:
        description: TBD
        items:
          type: string
        type: array
    type: object
  inframodel.OnpremInfra:
    properties:
      network:
        $ref: '#/definitions/inframodel.NetworkProperty'
      servers:
        items:
          $ref: '#/definitions/inframodel.ServerProperty'
        type: array
    required:
    - servers
    type: object
  inframodel.OsProperty:
    properties:
      id:
        example: ubuntu
        type: string
      idLike:
        example: debian
        type: string
      name:
        example: Ubuntu
        type: string
      prettyName:
        description: Pretty name
        example: Ubuntu 22.04.3 LTS
        type: string
      version:
        description: Full version string
        example: 22.04.3 LTS (Jammy Jellyfish)
        type: string
      versionCodename:
        example: jammy
        type: string
      versionId:
        example: "22.04"
        type: string
    required:
    - prettyName
    type: object
  inframodel.RouteProperty:
    properties:
      destination:
        description: Destination network, expressed in CIDR format
        type: string
      gateway:
        description: Gateway address to which packets are forwarded
        type: string
      interface:
        description: Network interface associated with the route
        type: string
      linkState:
        description: Link state of the route (e.g., UP, DOWN)
        type: string
      metric:
        description: Metric value indicating the priority of the route
        type: integer
      protocol:
        description: Protocol used to set the route (e.g., kernel, static)
        type: string
      scope:
        description: Scope of the route (e.g., global, link, host)
        type: string
      source:
        description: Optionally stores the source address (used for policy-based routing)
        type: string
    type: object
  inframodel.ServerProperty:
    properties:
      cpu:
        $ref: '#/definitions/inframodel.CpuProperty'
      dataDisks:
        items:
          $ref: '#/definitions/inframodel.DiskProperty'
        type: array
      hostname:
        type: string
      interfaces:
        items:
          $ref: '#/definitions/inframodel.NetworkInterfaceProperty'
        type: array
      memory:
        $ref: '#/definitions/inframodel.MemoryProperty'
      os:
        $ref: '#/definitions/inframodel.OsProperty'
      rootDisk:
        $ref: '#/definitions/inframodel.DiskProperty'
      routingTable:
        items:
          $ref: '#/definitions/inframodel.RouteProperty'
        type: array
    type: object
  model.TbMciDynamicReq:
    properties:
      description:
        example: Made in CB-TB
        type: string
      installMonAgent:
        default: "no"
        description: InstallMonAgent Option for CB-Dragonfly agent installation ([yes/no]
          default:yes)
        enum:
        - "yes"
        - "no"
        example: "no"
        type: string
      label:
        additionalProperties:
          type: string
        description: Label is for describing the object by keywords
        type: object
      name:
        example: mci01
        type: string
      systemLabel:
        description: SystemLabel is for describing the mci in a keyword (any string
          can be used) for special System purpose
        example: ""
        type: string
      vm:
        items:
          $ref: '#/definitions/model.TbVmDynamicReq'
        type: array
    required:
    - name
    - vm
    type: object
  model.TbVmDynamicReq:
    properties:
      commonImage:
        description: CommonImage is field for id of a image in common namespace
        example: ubuntu18.04
        type: string
      commonSpec:
        description: CommonSpec is field for id of a spec in common namespace
        example: aws+ap-northeast-2+t2.small
        type: string
      connectionName:
        description: |-
          if ConnectionName is given, the VM tries to use associtated credential.
          if not, it will use predefined ConnectionName in Spec objects
        type: string
      description:
        example: Description
        type: string
      label:
        additionalProperties:
          type: string
        description: Label is for describing the object by keywords
        type: object
      name:
        description: VM name or subGroup name if is (not empty) && (> 0). If it is
          a group, actual VM name will be generated with -N postfix.
        example: g1-1
        type: string
      rootDiskSize:
        default: default
        description: '"default", Integer (GB): ["50", ..., "1000"]'
        example: default, 30, 42, ...
        type: string
      rootDiskType:
        default: default
        description: '"", "default", "TYPE1", AWS: ["standard", "gp2", "gp3"], Azure:
          ["PremiumSSD", "StandardSSD", "StandardHDD"], GCP: ["pd-standard", "pd-balanced",
          "pd-ssd", "pd-extreme"], ALIBABA: ["cloud_efficiency", "cloud", "cloud_essd"],
          TENCENT: ["CLOUD_PREMIUM", "CLOUD_SSD"]'
        example: default, TYPE1, ...
        type: string
      subGroupSize:
        default: "1"
        description: if subGroupSize is (not empty) && (> 0), subGroup will be generated.
          VMs will be created accordingly.
        example: "3"
        type: string
      vmUserPassword:
        type: string
    required:
    - commonImage
    - commonSpec
    type: object
info:
  contact:
    email: contact-to-cloud-barista@googlegroups.com
    name: API Support
    url: http://cloud-barista.github.io
  description: CM-Damselfly REST API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: CM-Damselfly REST API
  version: latest
paths:
  /cloudmodel:
    get:
      consumes:
      - application/json
      description: Get a list of cloud models.
      produces:
      - application/json
      responses:
        "200":
          description: (sample) This is a list of models
          schema:
            $ref: '#/definitions/handler.GetCloudModelsResp'
        "404":
          description: model not found
          schema:
            type: object
      summary: Get a list of cloud models
      tags:
      - '[API] Cloud Migration User Models'
    post:
      consumes:
      - application/json
      description: Create a new cloud model with the given information.
      parameters:
      - description: model information
        in: body
        name: Model
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCloudModelReq'
      produces:
      - application/json
      responses:
        "201":
          description: (Sample) This is a sample description for success response
            in Swagger UI
          schema:
            $ref: '#/definitions/handler.CreateCloudModelResp'
        "400":
          description: Invalid Request
          schema:
            type: object
      summary: Create a new cloud model
      tags:
      - '[API] Cloud Migration User Models'
  /cloudmodel/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a cloud model with the given information.
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Model deletion successful
          schema:
            type: string
        "400":
          description: Invalid Request
          schema:
            type: object
        "404":
          description: Model Not Found
          schema:
            type: object
      summary: Delete a cloud model
      tags:
      - '[API] Cloud Migration User Models'
    get:
      consumes:
      - application/json
      description: Get a specific cloud model.
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: (Sample) a model
          schema:
            $ref: '#/definitions/handler.GetCloudModelResp'
        "404":
          description: model not found
          schema:
            type: object
      summary: Get a specific cloud model
      tags:
      - '[API] Cloud Migration User Models'
    put:
      consumes:
      - application/json
      description: Update a cloud model with the given information.
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      - description: Model information to update
        in: body
        name: Model
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateCloudModelReq'
      produces:
      - application/json
      responses:
        "201":
          description: (Sample) This is a sample description for success response
            in Swagger UI
          schema:
            $ref: '#/definitions/handler.UpdateCloudModelResp'
        "400":
          description: Invalid Request
          schema:
            type: object
      summary: Update a cloud model
      tags:
      - '[API] Cloud Migration User Models'
  /httpVersion:
    get:
      consumes:
      - application/json
      description: Checks and logs the HTTP version of the incoming request to the
        server console.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResHTTPVersion'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ResHTTPVersion'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ResHTTPVersion'
      summary: Check HTTP version of incoming request
      tags:
      - '[Admin] System management'
  /model/{isTargetModel}:
    get:
      consumes:
      - application/json
      description: Get a list of all user models.
      parameters:
      - description: Is TargetModel ?
        in: path
        name: isTargetModel
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: (sample) This is a list of models
          schema:
            $ref: '#/definitions/handler.GetModelsResp'
        "404":
          description: model not found
          schema:
            type: object
      summary: Get a list of all user models
      tags:
      - '[API] Migration User Models'
  /model/version:
    get:
      consumes:
      - application/json
      description: Get the versions of all models(schemata of on-premise/cloud migration
        models)
      produces:
      - application/json
      responses:
        "200":
          description: (sample) This is the versions of all models(schemata)
          schema:
            $ref: '#/definitions/handler.GetModelsVersionResp'
        "404":
          description: verson of models not found
          schema:
            type: object
      summary: Get the versions of all models(schemata of on-premise/cloud migration
        models)
      tags:
      - '[API] Migration Models'
  /onpremmodel:
    get:
      consumes:
      - application/json
      description: Get a list of on-premise models.
      produces:
      - application/json
      responses:
        "200":
          description: (sample) This is a list of models
          schema:
            $ref: '#/definitions/handler.GetOnPremModelsResp'
        "404":
          description: model not found
          schema:
            type: object
      summary: Get a list of on-premise models
      tags:
      - '[API] On-Premise Migration User Models'
    post:
      consumes:
      - application/json
      description: Create a new on-premise model with the given information.
      parameters:
      - description: model information
        in: body
        name: Model
        required: true
        schema:
          $ref: '#/definitions/handler.CreateOnPremModelReq'
      produces:
      - application/json
      responses:
        "201":
          description: (Sample) This is a sample description for success response
            in Swagger UI
          schema:
            $ref: '#/definitions/handler.CreateOnPremModelResp'
        "400":
          description: Invalid Request
          schema:
            type: object
      summary: Create a new on-premise model
      tags:
      - '[API] On-Premise Migration User Models'
  /onpremmodel/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a on-premise model with the given information.
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Model deletion successful
          schema:
            type: string
        "400":
          description: Invalid Request
          schema:
            type: object
        "404":
          description: Model Not Found
          schema:
            type: object
      summary: Delete a on-premise model
      tags:
      - '[API] On-Premise Migration User Models'
    get:
      consumes:
      - application/json
      description: Get a specific on-premise model.
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: (Sample) a model
          schema:
            $ref: '#/definitions/handler.GetOnPremModelResp'
        "404":
          description: model not found
          schema:
            type: object
      summary: Get a specific on-premise model
      tags:
      - '[API] On-Premise Migration User Models'
    put:
      consumes:
      - application/json
      description: Update a on-premise model with the given information.
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      - description: Model information to update
        in: body
        name: Model
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateOnPremModelReq'
      produces:
      - application/json
      responses:
        "201":
          description: (Sample) This is a sample description for success response
            in Swagger UI
          schema:
            $ref: '#/definitions/handler.UpdateOnPremModelResp'
        "400":
          description: Invalid Request
          schema:
            type: object
      summary: Update a on-premise model
      tags:
      - '[API] On-Premise Migration User Models'
  /readyz:
    get:
      consumes:
      - application/json
      description: Check Damselfly is ready
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResReadyz'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handler.ResReadyz'
      summary: Check Damselfly is ready
      tags:
      - '[Admin] System management'
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"